Towards multiple runs
Nomenclature changes, refactoring and moving some functions around
-renamed get_data_from_decimation_level_from_mth5 to get_data_from_mth5, since this
is not getting undecimated data.
-moved prototype_decimate into pipelines/time_series_helpers (for now)

[Issue(s): 118, 31]
